Clone originalSlides in cleanUpRows to keep registered events
When using unslick in combination with registered events (e.g. click on a slicked children element) even with jQuery .on() the events are removed. The reason is that empty() kills the events from the children before they are appended. By cloning them before, we keep the events. An example real world use case would be a modal registered on a slick element which doesn't work after unslicking anymore.